now supports numbers with decimals

+++ Calculator.java
@@ -5,10 +5,10 @@ public class Calculator {
         Scanner input = new Scanner(System.in); //sets up scanner
 
         System.out.println("Please enter two whole numbers\nFirst number?");
-        int firstNum = input.nextInt(); //saves input of first num
+        double firstNum = input.nextDouble(); //saves input of first num
 
         System.out.println("Second number?");
-        int secNum = input.nextInt(); //saves input of second num
+        double secNum = input.nextDouble(); //saves input of second num
 
         System.out.println("Please choose if you would like to add or subtract your numbers.\ntype \"a\" to add, \"s\" to subtract, \"m\" to multiply, or \"d\" to divide");
         input.nextLine();
@@ -25,7 +25,7 @@ public class Calculator {
         } else {
             System.out.println("That was not a valid entry, goodbye"); //if user hits wrong button
 
-        }
+        } 
 
     }
 }
